sql+python/requirements.txt
mysql-connector-python-8.2.0